move most of the files
