+++ internal/handler/msg_debugerror.go
@@ -16,12 +16,57 @@ package handler
 
 import (
 	"context"
+	"errors"
+	"strconv"
+	"strings"
 
-	"github.com/FerretDB/FerretDB/internal/handler/commoncommands"
+	"github.com/FerretDB/FerretDB/internal/handler/common"
+	"github.com/FerretDB/FerretDB/internal/handler/commonerrors"
+	"github.com/FerretDB/FerretDB/internal/types"
+	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
+	"github.com/FerretDB/FerretDB/internal/util/must"
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
 // MsgDebugError implements `debugError` command.
 func (h *Handler) MsgDebugError(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
-	return commoncommands.MsgDebugError(ctx, msg)
+	document, err := msg.Document()
+	if err != nil {
+		return nil, lazyerrors.Error(err)
+	}
+
+	expected, err := common.GetRequiredParam[string](document, document.Command())
+	if err != nil {
+		return nil, err
+	}
+
+	// check if parameter is an error code
+	if n, err := strconv.ParseInt(expected, 10, 32); err == nil {
+		errCode := commonerrors.ErrorCode(n)
+		return nil, errors.New(errCode.String())
+	}
+
+	switch {
+	case expected == "ok":
+		var reply wire.OpMsg
+
+		replyDoc := must.NotFail(types.NewDocument(
+			"ok", float64(1),
+		))
+
+		must.NoError(reply.SetSections(wire.OpMsgSection{
+			Documents: []*types.Document{replyDoc},
+		}))
+
+		return &reply, nil
+
+	case strings.HasPrefix(expected, "panic"):
+		panic("debugError " + expected)
+
+	case strings.HasPrefix(expected, "lazy"):
+		return nil, lazyerrors.New(expected)
+
+	default:
+		return nil, errors.New(expected)
+	}
 }
